refactoring for code review
